+++ orb/local_settings_staging.py
@@ -39,8 +39,7 @@ def modify(settings):
                                          ('audience', 'audience'), 
                                          ('geography', 'geography'), 
                                          ('language', 'language'), 
-                                         ('device', 'device'), 
-                                         ('license', 'license')]
+                                         ('device', 'device')]